// extensions/entrega-tienda-delivery/generated/api.ts
export type Maybe<T> = T | null;
// ...
export type Cart = HasMetafields & {
  __typename?: 'Cart';
  /**
   * The custom attributes associated with a cart to store additional information. Cart attributes
   * allow you to collect specific information from customers on the **Cart** page, such as order notes,
   * gift wrapping requests, or custom product details. Attributes are stored as key-value pairs.
   */
  attribute?: Maybe<Attribute>;
  /** The billing address associated with the cart. */
  billingAddress?: Maybe<MailingAddress>;
  /**
   * Information about the customer that's interacting with the cart. It includes details such as the
   * customer's email and phone number, and the total amount of money the customer has spent in the store.
   * This information helps personalize the checkout experience and ensures that accurate pricing and delivery options
   * are displayed to customers.
   */
  buyerIdentity?: Maybe<BuyerIdentity>;
  /**
   * A breakdown of the costs that the customer will pay at checkout. It includes the total amount,
   * the subtotal before taxes and duties, the tax amount, and duty charges.
   */
  cost: CartCost;
  /** The items in a cart that are eligible for fulfillment and can be delivered to the customer. */
  deliverableLines: Array<DeliverableCartLine>;
  /**
   * A collection of items that are grouped by shared delivery characteristics. Delivery groups streamline
   * fulfillment by organizing items that can be shipped together, based on the customer's
   * shipping address. For example, if a customer orders a t-shirt and a pair of shoes that can be shipped
   * together, then the items are included in the same delivery group.
   *
   * In the [Order Discount](https://shopify.dev/docs/api/functions/reference/order-discounts) and
   * [Product Discount](https://shopify.dev/docs/api/functions/reference/product-discounts) legacy APIs,
   * the `cart.deliveryGroups` input is always an empty array. This means you can't access delivery groups when
   * creating Order Discount or Product Discount Functions. If you need to apply discounts to shipping costs,
   * then use the [Discount Function API](https://shopify.dev/docs/api/functions/reference/discount)
   * instead.
   */
  deliveryGroups: Array<CartDeliveryGroup>;
  /** The discounts that have been applied to the cart. */
  discountApplications: Array<DiscountApplication>;
  /**
   * The items in a cart that the customer intends to purchase. A cart line is an entry in the
   * customer's cart that represents a single unit of a product variant. For example, if a customer adds two
   * different sizes of the same t-shirt to their cart, then each size is represented as a separate cart line.
   */
  lines: Array<CartLine>;
  /**
   * The additional fields on the **Cart** page that are required for international orders in specific countries,
   * such as customs information or tax identification numbers.
   */
  localizedFields: Array<LocalizedField>;
  /**
   * A [custom field](https://shopify.dev/docs/apps/build/custom-data) that stores additional information
   * about a Shopify resource, such as products, orders, and
   * [many more](https://shopify.dev/docs/api/admin-graphql/latest/enums/MetafieldOwnerType).
   * Using [metafields with Shopify Functions](https://shopify.dev/docs/apps/build/functions/input-output/metafields-for-input-queries)
   * enables you to customize the checkout experience.
   */
  metafield?: Maybe<Metafield>;
  /**
   * A purchase order number associated with the cart, often used for B2B transactions
   * to reference the buyer's internal purchase order.
   */
  poNumber?: Maybe<Scalars['String']['output']>;
  /** The physical location where a retail order is created or completed. */
  retailLocation?: Maybe<Location>;
};
// ...
export type CartDeliveryGroup = {
  __typename?: 'CartDeliveryGroup';
  /**
   * Information about items in a cart that a customer intends to purchase. A cart line is an entry in the
   * customer's cart that represents a single unit of a product variant. For example, if a customer adds two
   * different sizes of the same t-shirt to their cart, then each size is represented as a separate cart line.
   */
  cartLines: Array<CartLine>;
  /** The shipping or destination address associated with the delivery group. */
  deliveryAddress?: Maybe<MailingAddress>;
  /**
   * The delivery options available for the delivery group. Delivery options are the different ways that customers
   * can choose to have their orders shipped. Examples include express shipping or standard shipping.
   */
  deliveryOptions: Array<CartDeliveryOption>;
  /** The discounts that have been applied to the delivery group. */
  discountAllocations: Array<DiscountAllocation>;
  /** The type of merchandise in the delivery group. */
  groupType: CartDeliveryGroupType;
  /**
   * A [globally-unique ID](https://shopify.dev/docs/api/usage/gids)
   * for the delivery group.
   */
  id: Scalars['ID']['output'];
  /** Information about the delivery option that the customer has selected. */
  selectedDeliveryOption?: Maybe<CartDeliveryOption>;
};
// ...
export type CartDeliveryOptionsTransformRunResult = {
  /**
   * The ordered list of operations to apply to the list of
   * [delivery options](https://shopify.dev/docs/apps/build/checkout/delivery-shipping/delivery-options/build-function).
   */
  operations: Array<Operation>;
};
// ...
export enum DeliveryMethod {
  /** Local Delivery. */
  Local = 'LOCAL',
  /** None. */
  None = 'NONE',
  /** Shipping to a Pickup Point. */
  PickupPoint = 'PICKUP_POINT',
  /** Local Pickup. */
  PickUp = 'PICK_UP',
  /** Retail. */
  Retail = 'RETAIL',
  /** Shipping. */
  Shipping = 'SHIPPING'
}
// ...
export type Input = {
  __typename?: 'Input';
  /**
   * The cart where the Function is running. A cart contains the merchandise that a customer intends to purchase
   * and information about the customer, such as the customer's email address and phone number.
   */
  cart: Cart;
  /**
   * The backend logic that the Function is running to define how
   * [delivery options](https://shopify.dev/apps/build/checkout/delivery-shipping/delivery-options/build-function)
   * are sorted, hidden, or renamed. It includes the
   * [metafields](https://shopify.dev/docs/apps/build/custom-data)
   * that are associated with the customization.
   */
  deliveryCustomization: DeliveryCustomization;
  /**
   * The regional and language settings that determine how the Function
   * handles currency, numbers, dates, and other locale-specific values
   * during discount calculations. These settings are based on the store's configured
   * [localization practices](https://shopify.dev/docs/apps/build/functions/localization-practices-shopify-functions).
   */
  localization: Localization;
  /**
   * The exchange rate used to convert discounts between the shop's default
   * currency and the currency that displays to the customer during checkout.
   * For example, if a store operates in USD but a customer is viewing discounts in EUR,
   * then the presentment currency rate handles this conversion for accurate pricing.
   */
  presentmentCurrencyRate: Scalars['Decimal']['output'];
  /**
   * Information about the shop where the Function is running, including the shop's timezone
   * setting and associated [metafields](https://shopify.dev/docs/apps/build/custom-data).
   */
  shop: Shop;
};
// ...
export type MailingAddress = {
  __typename?: 'MailingAddress';
  /** The first line of the address. Typically the street address or PO Box number. */
  address1?: Maybe<Scalars['String']['output']>;
  /** The second line of the address. Typically the number of the apartment, suite, or unit. */
  address2?: Maybe<Scalars['String']['output']>;
  /** The name of the city, district, village, or town. */
  city?: Maybe<Scalars['String']['output']>;
  /** The name of the customer's company or organization. */
  company?: Maybe<Scalars['String']['output']>;
  /** The two-letter code for the country of the address. For example, US. */
  countryCode?: Maybe<CountryCode>;
  /** The first name of the customer. */
  firstName?: Maybe<Scalars['String']['output']>;
  /** The last name of the customer. */
  lastName?: Maybe<Scalars['String']['output']>;
  /** The approximate latitude of the address. */
  latitude?: Maybe<Scalars['Float']['output']>;
  /** The approximate longitude of the address. */
  longitude?: Maybe<Scalars['Float']['output']>;
  /**
   * The market of the address.
   * @deprecated This `market` field will be removed in a future version of the API.
   */
  market?: Maybe<Market>;
  /** The full name of the customer, based on firstName and lastName. */
  name?: Maybe<Scalars['String']['output']>;
  /** A unique phone number for the customer. Formatted using E.164 standard. For example, +16135551111. */
  phone?: Maybe<Scalars['String']['output']>;
  /** The alphanumeric code for the region. For example, ON. */
  provinceCode?: Maybe<Scalars['String']['output']>;
  /** The zip or postal code of the address. */
  zip?: Maybe<Scalars['String']['output']>;
};
// ...
export type MoneyV2 = {
  __typename?: 'MoneyV2';
  /**
   * A monetary value in decimal format, allowing for precise representation of cents or fractional
   * currency. For example, 12.99.
   */
  amount: Scalars['Decimal']['output'];
  /**
   * The three-letter currency code that represents a world currency used in a store. Currency codes
   * include standard [standard ISO 4217 codes](https://en.wikipedia.org/wiki/ISO_4217), legacy codes,
   * and non-standard codes. For example, USD.
   */
  currencyCode: CurrencyCode;
};
// ...
export type Operation =
  /** An operation that hides a delivery option from a list that's offered to customers at checkout. */
  { deliveryOptionHide: DeliveryOptionHideOperation; deliveryOptionMove?: never; deliveryOptionRename?: never; }
  |  /**
   * An operation that sorts a list of delivery options that are offered to customers at checkout.
   *
   * If you reorder shipping delivery options, then you are
   * [prohibited](https://shopify.dev/docs/apps/launch/app-requirements-checklist#prohibited-app-types).
   * from automatically selecting higher-priced delivery alternatives by default. The cheapest shipping delivery option
   * must always be the first option selected.
   */
  { deliveryOptionHide?: never; deliveryOptionMove: DeliveryOptionMoveOperation; deliveryOptionRename?: never; }
  |  /**
   * An operation that renames a delivery option that's offered to customers at checkout.
   *
   * The carrier name is automatically prepended to the delivery option title at checkout when using the
   * `DeliveryOptionRenameOperation` object, and can't be altered or omitted through the API. For example, if the carrier name
   * is **UPS** and the option is **Standard**, then you could change **UPS Standard** to **UPS Standard Shipping**,
   * but you couldn't change **UPS Standard** to **Standard Shipping**.
   */
  { deliveryOptionHide?: never; deliveryOptionMove?: never; deliveryOptionRename: DeliveryOptionRenameOperation; };
// ...
export type CartDeliveryOptionsTransformRunInput = { __typename?: 'Input', cart: { __typename?: 'Cart', billingAddress?: { __typename?: 'MailingAddress', city?: string | null, provinceCode?: string | null, zip?: string | null, countryCode?: CountryCode | null } | null, deliveryGroups: Array<{ __typename?: 'CartDeliveryGroup', deliveryAddress?: { __typename?: 'MailingAddress', city?: string | null, provinceCode?: string | null, zip?: string | null, countryCode?: CountryCode | null } | null, deliveryOptions: Array<{ __typename?: 'CartDeliveryOption', handle: any, title?: string | null, deliveryMethodType: DeliveryMethod, cost: { __typename?: 'MoneyV2', amount: any } }> }> }, shop: { __typename?: 'Shop', appEntregaConfig?: { __typename?: 'Metafield', jsonValue: any } | null, legacyEntregaConfig?: { __typename?: 'Metafield', jsonValue: any } | null } };